A bit more sync

+++ video-player/audio.py
@@ -42,6 +42,6 @@ class AudioPlayer:
 
 
 filename = sys.argv[1]
-start_time = int(sys.argv[2])
+start_time = float(sys.argv[2])
 audio_player = AudioPlayer(filename)
 audio_player.play(start_time)
